app/requirements.txt
requests==2.23.0
Flask==2.2.2
Flask-Cors==3.0.10
mysql-connector-python
